% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distPlots.R
\name{distPlots}
\alias{distPlots}
\title{Create Distribution Plots}
\usage{
distPlots(ppe, aggregate = TRUE, scaled = TRUE)
}
\arguments{
\item{ppe}{A specialized data object containing proteomics information.}

\item{aggregate}{A logical value indicating whether to aggregate the results (default is TRUE).}

\item{scaled}{A logical value indicating whether to use scaled values (default is TRUE).}
}
\value{
A ggplot object representing the histogram plot.
}
\description{
This function generates a histogram plot for the distribution of values from a specified data object.
It provides options for using scaled values and aggregating results.
}
