% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixWindows.R
\name{fixWindows}
\alias{fixWindows}
\title{Fix Missing Windows in Peptide Fragments}
\usage{
fixWindows(fragments)
}
\arguments{
\item{fragments}{A data frame containing peptide fragments with associated information.}
}
\value{
A data frame containing the fixed windows along with other associated information.
}
\description{
This function processes a data frame containing peptide fragments and attempts to fix missing window
information. It identifies missing windows and iteratively calls other functions to fix them.
}
\details{
Identify Missing Windows: The function identifies peptide fragments with missing window information and joins them with the original fragments data frame.
Fix Missing Windows (First Pass): The function iteratively calls the fixResidue function on each missing window to attempt to fix it. The fixed windows are then combined into a data frame.
Handle Still Missing Windows: The function identifies windows that are still missing after the first pass and attempts to fix them by retrieving protein sequences using the get_protein_sequence function and extracting windows using the getWindow function.
Determine Matches: The function determines matches by comparing various substrings of the stripped peptide sequence and window.
Fix Missing Windows (Second Pass): A second pass of fixing is performed, similar to the first pass, on the windows that are still missing.
Combine Fixed Windows: The function combines the fixed windows from both passes and filters those that match.
Combine with Original Fragments: The function combines the fixed windows with the original fragments that already had matching windows.
Return Result: The final result is returned, containing the fixed windows along with other associated information.
}
