% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_protein_sequence.R
\name{get_protein_sequence}
\alias{get_protein_sequence}
\title{Retrieve Protein Sequence from Swiss-Prot}
\usage{
get_protein_sequence(accession_number)
}
\arguments{
\item{accession_number}{A character string representing the accession number of the protein.}
}
\value{
A character string representing the protein sequence.
}
\description{
This function retrieves a protein sequence from the Swiss-Prot database using a given accession number.
It ensures that the Swiss-Prot bank is selected before performing the query.
}
