% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMotif.R
\name{getMotif}
\alias{getMotif}
\title{Retrieve Significant Motifs from Fitted Linear Model}
\usage{
getMotif(fit, sites = "STY", data, contrasts, LFC = 1.5)
}
\arguments{
\item{fit}{An object representing the fitted linear model (e.g., from lmFit).}

\item{sites}{A character string representing the sites of interest (default is "STY").}

\item{data}{A data frame containing information about the sequences, including 'id' and 'window' columns.}

\item{contrasts}{A character string or expression representing the contrasts to be used in the analysis.}

\item{LFC}{A numeric value representing the log fold change threshold for filtering the results (default is 1.5).}
}
\value{
An object representing the identified motifs based on the specified criteria.
}
\description{
This function retrieves significant motifs from a fitted linear model object based on specified
contrasts, log fold change (LFC) threshold, and sites of interest. It performs a motif search
using the specified foreground and background sequences and returns the identified motifs.
}
