% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPeptides.R
\name{getPeptides}
\alias{getPeptides}
\title{Retrieve Peptide Sequences}
\usage{
getPeptides(path_to_raw_data)
}
\arguments{
\item{path_to_raw_data}{A character string representing the path to the raw data file containing protein accessions.}
}
\value{
A data frame containing the final peptide sequences.
}
\description{
This function retrieves peptide sequences based on provided accessions. It uses various sources
including BioMart and custom functions to get the sequences, and handles missing values.
}
\details{
Read Accessions: The function reads a file containing protein accessions, splitting and unnesting them as needed.
Query Peptide Sequences: It queries the BioMart database to get peptide sequences for the given accessions using the query_biomart function and a specific filter attribute.
Load Peptide Sequences: The function loads previously saved peptide sequences from an RData file (this part is commented out in the code).
Find Missing Peptides: It finds the missing peptide sequences and attempts to recover them.
Query Other Accessions: The function performs additional queries using different filter attributes to get more peptide sequences.
Process Peptide Sequences: It processes the retrieved peptide sequences, removing any trailing stars.
Retrieve Remaining Missing Sequences: The function uses a loop to try to retrieve any remaining missing sequences using the get_protein_sequence function.
Combine Final Sequences: It combines all the retrieved sequences into a final result.
Return Final Sequences: The function returns the final peptide sequences.
}
