% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSite.R
\name{getSite}
\alias{getSite}
\title{Extract Phosphorylation Sites from Modified Peptide Sequence}
\usage{
getSite(modseq)
}
\arguments{
\item{modseq}{A character string representing the modified peptide sequence.}
}
\value{
A character string representing the positions of the phosphorylation sites, or NA if only one site is found.
}
\description{
This function extracts the positions of phosphorylation sites from a modified peptide sequence.
It handles various modifications and returns the positions of the phosphorylation sites.
}
\details{
Clean Input Sequence: The function removes specific modifications from the input sequence, such as "Carbamidomethyl (C)", "Acetyl (Protein N-term)", and "Oxidation (M)".
Split by Phosphorylation Sites: The cleaned sequence is split at each "\link{Phospho (STY)}" site, and the underscores are removed.
Calculate Positions: The function calculates the cumulative sum of the character lengths of the split strings, which represents the positions of the phosphorylation sites.
Return Positions: If only one phosphorylation site is found, the function returns NA. Otherwise, it returns the positions as a semicolon-separated string.
}
\examples{
getSite("_HSPAPPPDPGFPAPS[Phospho (STY)]PPPADSPSEGFS[Phospho (STY)]LK_")
getSite("_SKS[Phospho (STY)]PPKS[Phospho (STY)]PEEEGAVSS[Phospho (STY)]_")
}
