% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopHits.R
\name{getTopHits}
\alias{getTopHits}
\title{Retrieve Top Hits from Fitted Linear Model}
\usage{
getTopHits(fit, contrasts, lfc)
}
\arguments{
\item{fit}{An object representing the fitted linear model (e.g., from lmFit).}

\item{contrasts}{A character string or expression representing the contrasts to be used in the analysis.}

\item{lfc}{A numeric value representing the log fold change threshold for filtering the results.}
}
\value{
A data frame containing the top hits based on the specified criteria.
}
\description{
This function retrieves the top hits from a fitted linear model object based on specified
contrasts, p-value threshold, and log fold change (LFC) threshold. It is commonly used in
differential expression analysis.
}
