% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWindow.R
\name{getWindow}
\alias{getWindow}
\title{Extract Window of Peptide Sequence}
\usage{
getWindow(pos, peptide, window_size = 15)
}
\arguments{
\item{pos}{An integer representing the central position of the window.}

\item{peptide}{A character string representing the peptide sequence.}

\item{window_size}{An integer representing the size of the window (default is 15).}
}
\value{
A character string representing the extracted window of the peptide sequence.
}
\description{
This function extracts a specific window of a peptide sequence based on a given position
and window size. It handles edge cases by padding with underscores.
}
\details{
Calculate Window Bounds: The function calculates the left and right boundaries of the window based on the provided position and window size.
Handle Missing Values: If either the position or the peptide is missing (NA), the function sets the region to NA.
Extract Window: If the left and right sides of the window are within the peptide bounds, the function extracts the corresponding substring.
Handle Left Boundary: If the left side of the window is less than 1, the function pads the region with underscores.
Handle Right Boundary: If the right side of the window is greater than the length of the peptide, the function pads the region with underscores.
Return Region: The function returns the extracted region.
}
