% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeData.R
\name{normalizeData}
\alias{normalizeData}
\title{Normalize Phosphorylation Sites Data}
\usage{
normalizeData(sites, value_cutoff = 100, path)
}
\arguments{
\item{sites}{A data frame containing phosphorylation sites information.}

\item{value_cutoff}{A numeric value representing the cutoff for log2 transformation (default is 100).}

\item{path}{An optional character string representing the path to read the data from.}
}
\value{
A PhosphoExperiment object containing the normalized and processed data.
}
\description{
This function performs various preprocessing and normalization steps on a dataset containing
information about phosphorylation sites. It reads the data, applies transformations, filters,
imputes missing values, and scales the data.
}
