% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCplots.R
\name{QCplots}
\alias{QCplots}
\title{Create Quality Control (QC) Plots}
\usage{
QCplots(ppe)
}
\arguments{
\item{ppe}{A specialized data object containing proteomics information.}
}
\value{
A ggplot object representing the arranged QC plots (dendrogram and PCA).
}
\description{
This function creates quality control (QC) plots for a given data object. It generates
both dendrogram and principal component analysis (PCA) plots for the quantification and
scaled assays within the data object.
}
