% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_biomart.R
\name{query_biomart}
\alias{query_biomart}
\title{Query BioMart by Dividing Data into Chunks}
\usage{
query_biomart(data, filter_attribute)
}
\arguments{
\item{data}{A vector containing the data to be queried (e.g., UniProt accession IDs).}

\item{filter_attribute}{A filter attribute to be used in the BioMart query.}
}
\value{
A data frame containing the results of the BioMart query.
\dontrun{
  data <- c('P51587', 'Q8N726') # Add more UniProt accession IDs as needed
  result <- query_biomart(data, filter_attribute = "uniprotswissprot")
}
}
\description{
This function queries the BioMart database by dividing the input data into chunks of 200 elements each.
}
