% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testPPE.R
\name{testPPE}
\alias{testPPE}
\title{Perform Linear Model Fitting and Hypothesis Testing}
\usage{
testPPE(ppe, contrasts)
}
\arguments{
\item{ppe}{A specialized data object containing proteomics information.}

\item{contrasts}{A character string or expression representing the contrasts to be used in the analysis.}
}
\value{
An object representing the fitted model with contrasts and Empirical Bayes smoothing applied.
}
\description{
This function performs a linear model fitting and hypothesis testing on a given data object
using specified contrasts. It is designed for analyzing differential expression in proteomics data.
}
